nexusai-project/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
sentence-transformers==2.7.0
groq==0.9.0
httpx==0.27.0
PyMuPDF==1.24.5
pydantic==2.7.3
python-dotenv==1.0.1
numpy==1.26.4